+++ src/core/model/defs/gradient.hpp
@@ -24,6 +24,8 @@ public:
     QIcon reftarget_icon() const override;
     QString type_name_human() const override;
 
+    bool remove_if_unused(bool clean_lists) override;
+
 signals:
     void colors_changed(const QGradientStops&);
 };
@@ -59,7 +61,7 @@ public:
 
     static QString gradient_type_name(Type t);
 
-    void remove_if_unused() override;
+    bool remove_if_unused(bool clean_lists) override;
 
 private:
     std::vector<ReferenceTarget*> valid_refs() const;