+++ src/gui/widgets/dialogs/glaxnimate_window_p.hpp
@@ -119,8 +119,7 @@ public:
     void switch_tool(tools::Tool* tool);
     void switch_tool_action(QAction* action);
     void status_message(const QString& msg, int duration=5000);
-    void set_color_def_primary(model::BrushStyle* sty);
-    void set_color_def_secondary(model::BrushStyle* sty);
+    void set_color_def(model::BrushStyle* sty, bool secondary);
     QString get_open_image_file(const QString& title, const QString& dir);
     void set_brush_reference(model::BrushStyle* sty, bool secondary);
 
@@ -139,6 +138,7 @@ public:
     void group_shapes();
     void ungroup_shapes();
     void move_to();
+    void cleanup_document();
 
     void layer_new_layer();
     void layer_new_fill();