+++ src/gui/widgets/dialogs/gw_impl_document.cpp
@@ -492,58 +492,44 @@ QString GlaxnimateWindow::Private::drop_event_data(QDropEvent* event)
     return {};
 }
 
-void GlaxnimateWindow::Private::set_color_def_primary(model::BrushStyle* def)
+void GlaxnimateWindow::Private::set_color_def(model::BrushStyle* def, bool secondary)
 {
-    if ( auto target = ui.fill_style_widget->shape() )
+    model::Styler* target;
+    QString what;
+    if ( secondary )
+    {
+        target = ui.stroke_style_widget->shape();
+        what = tr("Fill");
+    }
+    else
+    {
+        target = ui.fill_style_widget->shape();
+        what = tr("Stroke");
+    }
+
+    if ( target )
     {
         auto old = target->use.get();
 
         if ( !def )
         {
-            command::UndoMacroGuard macro(tr("Unlink Fill Color"), current_document.get());
+            command::UndoMacroGuard macro(tr("Unlink %1 Color").arg(what), current_document.get());
             if ( auto col = qobject_cast<model::NamedColor*>(target->use.get()) )
                 target->color.set_undoable(col->color.get());
             target->use.set_undoable(QVariant::fromValue(def));
             if ( old )
-                old->remove_if_unused();
+                old->remove_if_unused(false);
         }
         else
         {
-            command::UndoMacroGuard macro(tr("Link Fill Color"), current_document.get());
+            command::UndoMacroGuard macro(tr("Link %1 Color").arg(what), current_document.get());
             target->use.set_undoable(QVariant::fromValue(def));
             if ( old )
-                old->remove_if_unused();
+                old->remove_if_unused(false);
         }
     }
 
     set_brush_reference(def, false);
-}
-
-void GlaxnimateWindow::Private::set_color_def_secondary(model::BrushStyle* def)
-{
-    if ( auto target = ui.stroke_style_widget->shape() )
-    {
-        auto old = target->use.get();
-
-        if ( !def )
-        {
-            command::UndoMacroGuard macro(tr("Unlink Stroke Color"), current_document.get());
-            if ( auto col = qobject_cast<model::NamedColor*>(target->use.get()) )
-                target->color.set_undoable(col->color.get());
-            target->use.set_undoable(QVariant::fromValue(def));
-            if ( old )
-                old->remove_if_unused();
-        }
-        else
-        {
-            command::UndoMacroGuard macro(tr("Link StrokeColor"), current_document.get());
-            target->use.set_undoable(QVariant::fromValue(def));
-            if ( old )
-                old->remove_if_unused();
-        }
-    }
-
-    set_brush_reference(def, true);
 }
 
 void GlaxnimateWindow::Private::set_brush_reference ( model::BrushStyle* sty, bool secondary )